Added authors
Make Post is a three modal process where the user inputs the content, then the author, and finally they get to preview the post before posting it. If the user does not enter a author, their username is used. The home page now shows the authors underneath the post and the post details does the same.