Added trending
Sorted by likes at the press of a button on the Home page and once pressed again, reverts back to the shuffled data. This is not the most likes from the database but is the same data as the shuffled data just sorted. The Home page fetches both the sorted and shuffled data which are in two separate arrays.